refactor: 미팅 이미지 수정 시 이미지파일이 들어오지 않았다면, 기존 이미지를 유지하도록 변경

+++ stargate-be/src/main/java/com/ssafy/stargate/model/service/MeetingServiceImpl.java
@@ -116,7 +116,7 @@ public class MeetingServiceImpl implements MeetingService {
      * @param dto       [MeetingCreateRequestDto] 입력받은 Meeting DTO
      * @param principal [Principal] 소속사 이메일이 포함된 객체
      * @return [MeetingResponseDto] 생성한 미팅 정보를 담은 DTO
-     * @throws CRUDException     데이터 CRUD 에러
+     * @throws CRUDException           데이터 CRUD 에러
      * @throws InputDataBlankException 빈 입력 데이터 에러
      */
     @Transactional
@@ -164,10 +164,11 @@ public class MeetingServiceImpl implements MeetingService {
 
     /**
      * 미팅 정보를 수정한다.
+     * 이미지파일이 없다면 전의 이미지를 변경하지 않는다.
      *
      * @param dto       [MeetingUpdateRequestDto] 입력받은 Meeting DTO
      * @param principal [Principal] 소속사 이메일이 포함된 객체
-     * @throws CRUDException     데이터 CRUD 에러
+     * @throws CRUDException           데이터 CRUD 에러
      * @throws InputDataBlankException 빈 입력 데이터 에러
      */
     @Transactional
@@ -189,7 +190,7 @@ public class MeetingServiceImpl implements MeetingService {
             meeting.setMeetingTime(dto.getMeetingTime());
             meeting.setPhotoNum(dto.getPhotoNum());
             meeting.setNotice(dto.getNotice());
-            meeting.setImage(newFilename);
+            meeting.setImage((newFilename != null) ? newFilename : meeting.getImage());
 
             List<MeetingMemberBridge> meetingMembers = meeting.getMeetingMembers();
             List<MeetingMemberBridge> newMeetingMembers = dtoToMeetingMemberList(meeting, MeetingCreateRequestDto.getParsedMeetingMembers(dto.getMeetingMembers()));
@@ -199,8 +200,10 @@ public class MeetingServiceImpl implements MeetingService {
             List<MeetingFUserBridge> newMeetingFUsers = dtoToMeetingFUserList(meeting, MeetingCreateRequestDto.getParsedMeetingFUsers(dto.getMeetingFUsers()));
             updateMeetingFUserList(meetingFUsers, newMeetingFUsers);
 
-            fileUtil.deleteFile(filePath, filename);
-        }  catch (InputDataBlankException inputDataBlankException) {
+            if (newFilename != null) {
+                fileUtil.deleteFile(filePath, filename);
+            }
+        } catch (InputDataBlankException inputDataBlankException) {
             throw inputDataBlankException; // 그대로 던져주기
         } catch (Exception e) {
             e.printStackTrace();
@@ -328,7 +331,7 @@ public class MeetingServiceImpl implements MeetingService {
      * @return List<MeetingMemberBridge> 미팅 멤버 엔티티 리스트
      * @throws NotFoundException             NotFoundException 데이터 찾기 실패 에러
      * @throws InputDataDuplicationException 미팅 멤버 리스트에 번호(id) 중복 발생 시 던져주는 에러
-     * @throws InputDataBlankException 빈 입력 데이터 에러
+     * @throws InputDataBlankException       빈 입력 데이터 에러
      */
     private List<MeetingMemberBridge> dtoToMeetingMemberList(Meeting meeting, List<Long> meetingMemberDtos) throws NotFoundException, InputDataDuplicationException, InputDataBlankException {
         if (meetingMemberDtos.size() < 1) {
@@ -360,7 +363,7 @@ public class MeetingServiceImpl implements MeetingService {
      * @param meetingFUserDtos [List<String>] 미팅 팬유저 이메일(id) 리스트
      * @return List<MeetingFUserBridge> 미팅 팬유저 엔티티 리스트
      * @throws InputDataDuplicationException 미팅 팬유저 리스트에 이메일(id) 중복 발생 시 던져주는 에러
-     * @throws InputDataBlankException 빈 입력 데이터 에러
+     * @throws InputDataBlankException       빈 입력 데이터 에러
      */
     private List<MeetingFUserBridge> dtoToMeetingFUserList(Meeting meeting, List<String> meetingFUserDtos) throws InputDataDuplicationException, InputDataBlankException {
         if (meetingFUserDtos.size() < 1) {
